Reading the values from Firebase in order
This change uses the right pattern to read the values for the metric in
timestamp order and actually limit the reading to the requested 100 items. This
change has the side effect that multiple messages will be sent to the watch as
new values are aquired. The guarantee is that a max of 100 values will be
delivered to the watch at any given moment.